Split some actions Pt 1 (#3773)
Moving LoadingScreen and CaseSelect Actions
