+++ client/app/reader/DocumentsTable.jsx
@@ -16,7 +16,8 @@ import DropdownFilter from './DropdownFilter';
 import { bindActionCreators } from 'redux';
 import Highlight from '../components/Highlight';
 import { setDocListScrollPosition, changeSortState, clearTagFilters, clearCategoryFilters,
-  setTagFilter, setCategoryFilter, toggleDropdownFilterVisibility } from './actions';
+  setTagFilter, setCategoryFilter, toggleDropdownFilterVisibility
+} from '../reader/actions';
 import { getAnnotationsPerDocument } from './selectors';
 import {
   SortArrowUp, SortArrowDown, DoubleArrow } from '../components/RenderFunctions';
