+++ client/app/reader/constants.js
@@ -71,7 +71,7 @@ export const SET_PDF_DOCUMENT = 'SET_PDF_DOCUMENT';
 export const CLEAR_PDF_PAGE = 'CLEAR_PDF_PAGE';
 export const ROTATE_PDF_DOCUMENT = 'ROTATE_PDF_DOCUMENT';
 export const CLEAR_PDF_DOCUMENT = 'CLEAR_PDF_DOCUMENT';
-export const GET_DCOUMENT_TEXT = 'GET_DCOUMENT_TEXT';
+export const GET_DOCUMENT_TEXT = 'GET_DCOUMENT_TEXT';
 export const ZERO_SEARCH_INDEX = 'ZERO_SEARCH_INDEX';
 export const UPDATE_SEARCH_INDEX = 'UPDATE_SEARCH_INDEX';
